docs: add command to remove background color
I added the command to remove the background from the NormalNC highlight that was missing. This highlight takes care of the background color when the main window is not in focus. For example, when we open the telescope float window or when we open the nvim-tree and the main window loses focus.